Condominio de volta a ser String

+++ src/main/java/com/br/Kodominio/dao/ICondominio.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @EnableJpaRepositories
-public interface ICondominio extends CrudRepository<Condominio, Integer> {
+public interface ICondominio extends CrudRepository<Condominio, Long> {
 
 }
